throttled service warnings and a bit of a tidy up
