+++ package.json
@@ -7,6 +7,7 @@
     "release": "npm-prepublish --verbose && npm publish && git checkout package.json"
   },
   "dependencies": {
+    "debounce": "^1.0.0",
     "debug": "^2.2.0",
     "es6-promise": "^2.0.1",
     "es6-symbol": "^2.0.1",
